+++ src/lib_gfx/vm_core.h
@@ -236,12 +236,12 @@ uint64_t vm_u64_get_next_tick(
 int32_t vm_i32_get_data( 
     uint8_t *stream,
     vm_state_t *state,
-    uint8_t addr );
+    uint16_t addr );
 
 void vm_v_get_data_multi( 
     uint8_t *stream,
     vm_state_t *state,
-    uint8_t addr, 
+    uint16_t addr, 
     uint16_t len,
     int32_t *dest );
 
@@ -252,7 +252,7 @@ int32_t* vm_i32p_get_data_ptr(
 void vm_v_set_data( 
     uint8_t *stream,
     vm_state_t *state,
-    uint8_t addr, 
+    uint16_t addr, 
     int32_t data );
 
 int8_t vm_i8_load_program(
